refactor: NormalStatsTableViewからNormalStatsTableRowを切り出し
