+++ RugStats/View/Stats/NormalStatsTableView.swift
@@ -7,22 +7,7 @@
 
 import SwiftUI
 
-struct NormalStatsTableView: View {
-    // DateFormatter はプロパティの初期化内で設定する
-    let dateFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateStyle = .none
-        formatter.timeStyle = .medium
-        return formatter
-    }()
-    
-    // ゲームクロックをフォーマットする関数
-    private func formatTimeInterval(_ time: TimeInterval) -> String {
-        let minutes = Int(time) / 60
-        let seconds = Int(time) % 60
-        return String(format: "%02d:%02d", minutes, seconds)
-    }
-    
+struct NormalStatsTableView: View {    
     @Binding var timeline: [TimelineItem]
     
     var body: some View {
@@ -52,18 +37,7 @@ struct NormalStatsTableView: View {
                     VStack(alignment: .leading){
                         // データ部分を作成
                         ForEach(timeline) { item in
-                            HStack {
-                                Text(item.actorName)
-                                    .frame(width: 100, alignment: .leading)
-                                Text(item.actionName)
-                                    .frame(width: 100, alignment: .leading)
-                                Text(item.actionLabels.joined(separator: "、"))
-                                    .frame(width: 100, alignment: .leading)
-                                Text(formatTimeInterval(item.gameClock))
-                                    .frame(width: 50, alignment: .leading)
-                                Text(dateFormatter.string(from: item.timestamp))
-                                    .frame(width: 100, alignment: .leading)
-                            }
+                            NormalStatsTableRow(item: item)
                             Divider()
                         }
                     }
